Add canceld/not going info to card, Also added when platform is changed that the platform is displayed in red
Added info when train is cancelled using the same location as is used when train delays. if train is cancelled this is now shown next to the original departure time. When train is cancelled there is no platform so changed platform to -.

Also added that when a platform is changed (platform_planned is not platform_actual), that the platform text becomes red.

It works fine on my own installation. hopefully you can consider to add this to your card for everyone using it.